+++ framework/contracts/native/ibc-client/src/commands.rs
@@ -17,7 +17,7 @@ use abstract_core::{
 };
 use abstract_sdk::{
     core::{
-        ibc_client::state::{ACCOUNTS, ADMIN, CONFIG},
+        ibc_client::state::{ACCOUNTS, CONFIG},
         ibc_host::{HostAction, InternalAction},
         objects::{ans_host::AnsHost, version_control::VersionControlContract, ChannelEntry},
         ICS20,
@@ -41,7 +41,7 @@ pub fn execute_update_config(
     new_version_control: Option<String>,
 ) -> IbcClientResult {
     // auth check
-    ADMIN.assert_admin(deps.as_ref(), &info.sender)?;
+    cw_ownable::assert_owner(deps.storage, &info.sender)?;
     let mut cfg = CONFIG.load(deps.storage)?;
 
     if let Some(ans_host) = new_ans_host {
@@ -73,7 +73,7 @@ pub fn execute_register_infrastructure(
 ) -> IbcClientResult {
     let host_chain = ChainName::from_str(&host_chain)?;
     // auth check
-    ADMIN.assert_admin(deps.as_ref(), &info.sender)?;
+    cw_ownable::assert_owner(deps.storage, &info.sender)?;
 
     let note = deps.api.addr_validate(&note)?;
     // Can't allow if it already exists
@@ -119,7 +119,7 @@ pub fn execute_remove_host(
 ) -> IbcClientResult {
     let host_chain = ChainName::from_str(&host_chain)?;
     // auth check
-    ADMIN.assert_admin(deps.as_ref(), &info.sender)?;
+    cw_ownable::assert_owner(deps.storage, &info.sender)?;
 
     if let Some(ibc_infra) = IBC_INFRA.may_load(deps.storage, &host_chain)? {
         REVERSE_POLYTONE_NOTE.remove(deps.storage, &ibc_infra.polytone_note);