+++ framework/contracts/native/ibc-client/src/error.rs
@@ -4,7 +4,6 @@ use abstract_core::{
 };
 use abstract_sdk::AbstractSdkError;
 use cosmwasm_std::StdError;
-use cw_controllers::AdminError;
 use polytone::callbacks::CallbackMessage;
 use thiserror::Error;
 
@@ -20,7 +19,7 @@ pub enum IbcClientError {
     AbstractSdk(#[from] AbstractSdkError),
 
     #[error("{0}")]
-    Admin(#[from] AdminError),
+    Ownership(#[from] cw_ownable::OwnershipError),
 
     #[error("{0}")]
     VersionControlError(#[from] VersionControlError),